+++ MasterOfHair/MasterOfHair/HomePage/jinkahuiyuan/JinkahuiyuanViewController.m
@@ -236,7 +236,7 @@
                            [SVProgressHUD showSuccessWithStatus:@"支付成功~" maskType:SVProgressHUDMaskTypeBlack];
                        } else {
                            // 支付失败或取消
-                           NSLog(@"Error: code=%lu msg=%@", error.code, [error getMsg]);
+                           NSLog(@"Error: code=%lu msg=%@", (unsigned long)error.code, [error getMsg]);
                            [SVProgressHUD showErrorWithStatus:@"支付失败~" maskType:SVProgressHUDMaskTypeBlack];
                        }
                    }];
